migrate to target dev/1.0.0, remove unwraps
cleanup

+++ src/lib.rs
@@ -20,16 +20,20 @@ use std::path::PathBuf;
 use std::time::Duration;
 use tracing::{debug, error, trace, warn};
 use uhlc::NTP64;
-use zenoh::buffers::{reader::HasReader, writer::HasWriter};
-use zenoh::prelude::*;
-use zenoh::properties::Properties;
-use zenoh::time::{new_reception_timestamp, Timestamp};
+use zenoh::encoding::Encoding;
+use zenoh::time::{new_timestamp, Timestamp};
+use zenoh::value::Value;
 use zenoh::Result as ZResult;
 use zenoh_backend_traits::config::{StorageConfig, VolumeConfig};
 use zenoh_backend_traits::*;
+use zenoh_buffers::buffer::SplitBuffer;
+use zenoh_buffers::ZBuf;
+use zenoh_buffers::{reader::HasReader, writer::HasWriter};
 use zenoh_codec::{RCodec, WCodec, Zenoh080};
 use zenoh_core::{bail, zerror};
+use zenoh_keyexpr::OwnedKeyExpr;
 use zenoh_plugin_trait::{plugin_long_version, plugin_version, Plugin};
+use zenoh_protocol::core::Properties;
 use zenoh_util::zenoh_home;
 
 /// The environement variable used to configure the root of all storages managed by this RocksdbBackend.
@@ -89,8 +93,8 @@ impl Plugin for RocksDbBackend {
             dir
         };
         let mut properties = Properties::default();
-        properties.insert("root".into(), root.to_string_lossy().into());
-        properties.insert("version".into(), Self::PLUGIN_VERSION.into());
+        properties.insert::<String, String>("root".into(), root.to_string_lossy().into());
+        properties.insert::<String, String>("version".into(), Self::PLUGIN_VERSION.into());
 
         let admin_status = HashMap::from(properties)
             .into_iter()
@@ -197,14 +201,6 @@ impl Volume for RocksdbVolume {
             db,
         }))
     }
-
-    fn incoming_data_interceptor(&self) -> Option<Arc<dyn Fn(Sample) -> Sample + Send + Sync>> {
-        None
-    }
-
-    fn outgoing_data_interceptor(&self) -> Option<Arc<dyn Fn(Sample) -> Sample + Send + Sync>> {
-        None
-    }
 }
 
 struct RocksdbStorage {
@@ -230,7 +226,10 @@ impl Storage for RocksdbStorage {
     ) -> ZResult<StorageInsertionResult> {
         // Get lock on DB
         let db_cell = self.db.lock().await;
-        let db = db_cell.as_ref().unwrap();
+
+        let db = db_cell
+            .as_ref()
+            .ok_or(zerror!("Could not get DB ref in put"))?;
 
         // Store the sample
         debug!(
@@ -254,7 +253,9 @@ impl Storage for RocksdbStorage {
     ) -> ZResult<StorageInsertionResult> {
         // Get lock on DB
         let db_cell = self.db.lock().await;
-        let db = db_cell.as_ref().unwrap();
+        let db = db_cell
+            .as_ref()
+            .ok_or(zerror!("Could not get DB ref in delete"))?;
 
         debug!("Deleting key: {:?}", key);
         if !self.read_only {
@@ -274,7 +275,9 @@ impl Storage for RocksdbStorage {
     ) -> ZResult<Vec<StoredData>> {
         // Get lock on DB
         let db_cell = self.db.lock().await;
-        let db = db_cell.as_ref().unwrap();
+        let db = db_cell
+            .as_ref()
+            .ok_or(zerror!("Could not get DB ref in get"))?;
 
         // Get the matching key/value
         debug!("getting key `{:?}` with parameters `{}`", key, _parameters);
@@ -289,10 +292,16 @@ impl Storage for RocksdbStorage {
         let mut result = Vec::new();
 
         let db_cell = self.db.lock().await;
-        let db = db_cell.as_ref().unwrap();
+        let db = db_cell
+            .as_ref()
+            .ok_or(zerror!("Could not get DB ref in get_all_entries"))?;
+
+        let cf_handle = db.cf_handle(CF_DATA_INFO).ok_or(zerror!(
+            "Option for ColumFamily {CF_DATA_INFO} was None, cancel get_all_entries"
+        ))?;
 
         // Iterate over DATA_INFO Column Family to avoid loading payloads
-        for item in db.prefix_iterator_cf(db.cf_handle(CF_DATA_INFO).unwrap(), "") {
+        for item in db.prefix_iterator_cf(cf_handle, "") {
             let (key, buf) = item.map_err(|e| zerror!("{}", e))?;
             let key_str = String::from_utf8_lossy(&key);
             let res_ke = if key_str == NONE_KEY {
@@ -323,40 +332,43 @@ impl Drop for RocksdbStorage {
             // Get lock on DB and take DB so we can drop it before destroying it
             // (avoiding RocksDB lock to be taken twice)
             let mut db_cell = self.db.lock().await;
-            let db = db_cell.take().unwrap();
 
-            // Flush all
-            if let Err(err) = db.flush() {
-                warn!("Closing Rocksdb storage, flush failed: {}", err);
-            }
+            if let Some(db) = db_cell.take() {
+                // Flush all
+                if let Err(err) = db.flush() {
+                    warn!("Closing Rocksdb storage, flush failed: {}", err);
+                }
 
-            // copy path for later use after DB is dropped
-            let path = db.path().to_path_buf();
+                // copy path for later use after DB is dropped
+                let path = db.path().to_path_buf();
 
-            // drop DB, releasing RocksDB lock
-            drop(db);
+                // drop DB, releasing RocksDB lock
+                drop(db);
 
-            match self.on_closure {
-                OnClosure::DestroyDB => {
-                    debug!(
-                        "Close Rocksdb storage, destroying database {}",
-                        path.display()
-                    );
-                    if let Err(err) = DB::destroy(&Options::default(), &path) {
-                        error!(
-                            "Failed to destroy Rocksdb database '{}' : {}",
-                            path.display(),
-                            err
+                match self.on_closure {
+                    OnClosure::DestroyDB => {
+                        debug!(
+                            "Close Rocksdb storage, destroying database {}",
+                            path.display()
+                        );
+                        if let Err(err) = DB::destroy(&Options::default(), &path) {
+                            error!(
+                                "Failed to destroy Rocksdb database '{}' : {}",
+                                path.display(),
+                                err
+                            );
+                        }
+                    }
+                    OnClosure::DoNothing => {
+                        debug!(
+                            "Close Rocksdb storage, keeping database {} as it is",
+                            path.display()
                         );
                     }
                 }
-                OnClosure::DoNothing => {
-                    debug!(
-                        "Close Rocksdb storage, keeping database {} as it is",
-                        path.display()
-                    );
-                }
-            }
+            } else {
+                warn!("Tried Dropping DB connection, however D Connection internally was None, Continuing");
+            };
         });
     }
 }
@@ -368,7 +380,8 @@ fn put_kv(
     timestamp: Timestamp,
 ) -> ZResult<StorageInsertionResult> {
     trace!("Put key {:?} in {:?}", key, db);
-    let data_info = encode_data_info(&value.encoding, &timestamp, false)?;
+
+    let data_info = encode_data_info(value.encoding().clone(), &timestamp, false)?;
 
     let key = match key {
         Some(k) => k.to_string(),
@@ -377,11 +390,19 @@ fn put_kv(
     // Write content and encoding+timestamp in different Column Families
     let mut batch = WriteBatch::default();
     batch.put_cf(
-        db.cf_handle(CF_PAYLOADS).unwrap(),
+        db.cf_handle(CF_PAYLOADS).ok_or(zerror!(
+            "Option for ColumFamily {CF_PAYLOADS} was None, cancel put_kv"
+        ))?,
         &key,
-        value.payload.contiguous(),
+        ZBuf::from(value.payload()).contiguous(),
     );
-    batch.put_cf(db.cf_handle(CF_DATA_INFO).unwrap(), &key, data_info);
+    batch.put_cf(
+        db.cf_handle(CF_DATA_INFO).ok_or(zerror!(
+            "Option for ColumFamily {CF_DATA_INFO} was None, cancel put_kv"
+        ))?,
+        &key,
+        data_info,
+    );
 
     match db.write(batch) {
         Ok(()) => Ok(StorageInsertionResult::Inserted),
@@ -398,8 +419,19 @@ fn delete_kv(db: &DB, key: Option<OwnedKeyExpr>) -> ZResult<StorageInsertionResu
         Some(k) => k.to_string(),
         None => NONE_KEY.to_string(),
     };
-    batch.delete_cf(db.cf_handle(CF_PAYLOADS).unwrap(), &key);
-    batch.delete_cf(db.cf_handle(CF_DATA_INFO).unwrap(), &key);
+
+    if let Some(cf_payloads) = db.cf_handle(CF_PAYLOADS) {
+        batch.delete_cf(cf_payloads, &key);
+    } else {
+        warn!("Option for ColumFamily {CF_PAYLOADS} was None, continue with delete_kv, ignoring CF_PAYLOADS");
+    }
+
+    if let Some(cf_data_info) = db.cf_handle(CF_DATA_INFO) {
+        batch.delete_cf(cf_data_info, &key);
+    } else {
+        warn!("Option for ColumFamily {CF_DATA_INFO} was None, continue with delete_kv, ignoring CF_DATA_INFO");
+    }
+
     match db.write(batch) {
         Ok(()) => Ok(StorageInsertionResult::Deleted),
         Err(e) => Err(rocksdb_err_to_zerr(e)),
@@ -413,37 +445,54 @@ fn get_kv(db: &DB, key: Option<OwnedKeyExpr>) -> ZResult<Option<(Value, Timestam
         Some(k) => k.to_string(),
         None => NONE_KEY.to_string(),
     };
-    match (
-        db.get_cf(db.cf_handle(CF_PAYLOADS).unwrap(), &key),
-        db.get_cf(db.cf_handle(CF_DATA_INFO).unwrap(), &key),
-    ) {
+
+    let (cf_payloads, cf_data_info) = match (db.cf_handle(CF_PAYLOADS), db.cf_handle(CF_DATA_INFO))
+    {
+        (None, None) => {
+            bail!("Option for ColumFamily {CF_PAYLOADS} and {CF_DATA_INFO} were both 'None'")
+        }
+        (None, Some(_)) => bail!("Option for ColumFamily and {CF_DATA_INFO} were both 'None'"),
+        (Some(_), None) => bail!("Option for ColumFamily and {CF_PAYLOADS} were both 'None'"),
+        (Some(cf_payloads), Some(cf_data_info)) => (cf_payloads, cf_data_info),
+    };
+
+    match (db.get_cf(cf_payloads, &key), db.get_cf(cf_data_info, &key)) {
         (Ok(Some(payload)), Ok(Some(info))) => {
             trace!("first ok");
             let (encoding, timestamp, deleted) = decode_data_info(&info)?;
+
             if deleted {
                 Ok(None)
             } else {
-                Ok(Some((
-                    Value::new(payload.into()).encoding(encoding),
-                    timestamp,
-                )))
+                Ok(Some((Value::new(payload, encoding), timestamp)))
             }
         }
         (Ok(Some(payload)), Ok(None)) => {
             trace!("second ok");
             // Only the payload is present in DB!
             // Possibly legacy data. Consider as encoding as APP_OCTET_STREAM and create timestamp from now()
-            Ok(Some((
-                Value::new(payload.into()).encoding(KnownEncoding::AppOctetStream.into()),
-                new_reception_timestamp(),
-            )))
+            match std::num::NonZeroU128::new(1u128)
+                .map(new_timestamp)
+                .map(|timestamp| {
+                    (
+                        Value::new(payload, Encoding::APPLICATION_OCTET_STREAM),
+                        timestamp,
+                    )
+                })
+                .map(|x| Ok(Some(x)))
+            {
+                Some(x) => x,
+                None => {
+                    bail!("Error Creating Timestamp ID for 0x01");
+                }
+            }
         }
         (Ok(None), _) => Ok(None),
         (Err(err), _) | (_, Err(err)) => Err(rocksdb_err_to_zerr(err)),
     }
 }
 
-fn encode_data_info(encoding: &Encoding, timestamp: &Timestamp, deleted: bool) -> ZResult<Vec<u8>> {
+fn encode_data_info(encoding: Encoding, timestamp: &Timestamp, deleted: bool) -> ZResult<Vec<u8>> {
     let codec = Zenoh080::new();
     let mut result = vec![];
     let mut writer = result.writer();
@@ -455,27 +504,33 @@ fn encode_data_info(encoding: &Encoding, timestamp: &Timestamp, deleted: bool) -
     codec
         .write(&mut writer, deleted as u8)
         .map_err(|_| zerror!("Failed to encode data-info (deleted)"))?;
+
     codec
-        .write(&mut writer, encoding)
+        .write(&mut writer, &zenoh_protocol::core::Encoding::from(encoding))
         .map_err(|_| zerror!("Failed to encode data-info (encoding)"))?;
     Ok(result)
 }
 
 fn decode_data_info(buf: &[u8]) -> ZResult<(Encoding, Timestamp, bool)> {
     let codec = Zenoh080::new();
+
     let mut reader = buf.reader();
+
     let timestamp: Timestamp = codec
         .read(&mut reader)
         .map_err(|_| zerror!("Failed to decode data-info (timestamp)"))?;
+
     let deleted: u8 = codec
         .read(&mut reader)
         .map_err(|_| zerror!("Failed to decode data-info (deleted)"))?;
-    let encoding: Encoding = codec
+
+    let encoding: zenoh_protocol::core::Encoding = codec
         .read(&mut reader)
         .map_err(|_| zerror!("Failed to decode data-info (timestamp)"))?;
+
     let deleted = deleted != 0;
 
-    Ok((encoding, timestamp, deleted))
+    Ok((encoding.into(), timestamp, deleted))
 }
 
 fn rocksdb_err_to_zerr(err: rocksdb::Error) -> zenoh_core::Error {
